Issue #35 Installing LIO on GHC 7.4
https://github.com/scslab/lio/issues/35

Turns out ubuntu is still on old version of GHC so our version of LIO
doesn't compile on machines with GHC 7.4. This patch works around the
uglyness of the API changes.

I am not very happy with this patch since it's less clean than what we
had before, so once we no longer wish to support old version of GHC we
should revert this.